ensure deserialization on storage event in decorator
use service.get instead of ev.newValue to ensure that value is deserialized. otherwise a storage event triggered from another tab leads to an endless recursion because the value is used as a bare string which gets serialized, adding an additional pair of quotation marks every time, triggering events back and forth between the two tabs.